scraping/pyproject.toml
[project]
name = "fastboard-scraping"
version = "0.1.0"
description = "LinkedIn profile scraping service for FastboardAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "playwright>=1.49.0",
    "duckduckgo-search>=7.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
scraping = "scraping.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraping"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
